add cfg
