--- components/resource_metering/tests/mod.rs
@@ -1,4 +1,0 @@
-// Copyright 2021 TiKV Project Authors. Licensed under Apache-2.0.
-
-#[cfg(target_os = "linux")]
-mod cpu_recorder_linux;
